chore: Fix Python linting
